+++ mopidy/core/playback.py
@@ -289,7 +289,7 @@ class PlaybackController(object):
             # For testing
             self._volume = volume
 
-        self._trigger_volume_changed()
+        self._trigger_volume_changed(volume)
 
     volume = property(get_volume, set_volume)
     """Volume as int in range [0..100] or :class:`None`"""
@@ -517,9 +517,9 @@ class PlaybackController(object):
         logger.debug('Triggering options changed event')
         listener.CoreListener.send('options_changed')
 
-    def _trigger_volume_changed(self):
+    def _trigger_volume_changed(self, volume):
         logger.debug('Triggering volume changed event')
-        listener.CoreListener.send('volume_changed')
+        listener.CoreListener.send('volume_changed', volume=volume)
 
     def _trigger_seeked(self, time_position):
         logger.debug('Triggering seeked event')